feat: add function send_datapoint
Function takes measurement values as arrays and a int value for the time
and sends each array/value to the serial output. The arrays are send as
a whole in form of a series of bytes. The reciever should be able to
decode the entrie array rather than individual values within it.