feat (datalogger): add functionality for sending of duration
Pi keeps sending string '2' as a signal that it is ready to recieve the duration of the current cycle.
After that Arduino send duration and Pi stores it in datapoint list.
Datapoint list is now complet for this Cycle and it is ready to be written to the csv file.