refactor: remove redudant code part
The code for letting the system wait for the current cycle timeframe to end
is redudant.
By the time the door closes the whileloop tracking the cycle time only
keeps reading the
sensors. (test pending)

In other words: Once the door closes all other paths except those reading the
sensors are blocked. Only the sensor reading happens until the timeframes end
is reaached and a new cycle starts.