fix: change type of variable
The variable to track the cycle duration has been defined as an int.
Thus the variable yielded negative values when its range limit was reached.

Setting the type to unsigned long solved the problem. The point in time
when a new cycle starts can now be track as indented.

IMPORTANT NOTE!!
The sippet now works as it should:
the specified number of cycles is performed and each cycle last for 15 sec.
Waiting for the timeframe to pass when the door is closed again.